chore: reparent mock nostr collab relay
